@@ -22,32 +22,32 @@
 		"test:unit": "vitest"
 	},
 	"peerDependencies": {
-		"svelte": "^4.2.19"
+		"svelte": "^5.0.0"
 	},
 	"devDependencies": {
-		"@playwright/test": "1.47.2",
-		"@sveltejs/adapter-cloudflare": "4.7.4",
-		"@sveltejs/kit": "2.5.28",
-		"@sveltejs/package": "2.3.7",
-		"@sveltejs/vite-plugin-svelte": "3.1.2",
-		"@types/eslint": "8.56.12",
-		"beercss": "3.6.13",
-		"eslint": "8.57.1",
+		"@playwright/test": "^1.48.2",
+		"@sveltejs/adapter-cloudflare": "^4.7.4",
+		"@sveltejs/kit": "^2.8.1",
+		"@sveltejs/package": "^2.3.7",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@types/eslint": "^9.6.1",
+		"beercss": "^3.7.12",
+		"eslint": "^9.14.0",
 		"eslint-config-prettier": "9.1.0",
-		"eslint-plugin-svelte": "2.46.0",
-		"marked": "12.0.2",
-		"marked-base-url": "1.1.5",
-		"marked-gfm-heading-id": "3.2.0",
-		"prettier": "3.3.3",
-		"prettier-plugin-svelte": "3.2.7",
-		"publint": "0.2.12",
-		"svelte": "4.2.19",
-		"svelte-check": "3.8.6",
-		"svelte-infinite-loading": "1.3.8",
-		"tslib": "2.6.3",
-		"typescript": "5.3.3",
-		"vite": "5.1.8",
-		"vitest": "1.2.2"
+		"eslint-plugin-svelte": "^2.46.0",
+		"marked": "^15.0.0",
+		"marked-base-url": "^1.1.6",
+		"marked-gfm-heading-id": "^4.1.1",
+		"prettier": "^3.3.3",
+		"prettier-plugin-svelte": "^3.2.8",
+		"publint": "^0.2.12",
+		"svelte": "^5.1.16",
+		"svelte-check": "^4.0.7",
+		"svelte-infinite-loading": "^1.4.0",
+		"tslib": "^2.8.1",
+		"typescript": "^5.6.3",
+		"vite": "^5.4.11",
+		"vitest": "^2.1.5"
 	},
 	"files": [
 		"dist",
